refactor(RecommendedScheduleSortStandard): 추천 일정의 정렬 기준을 정의하는 enum 이름 변경
- Enum에서 Comparator 구현 분리
