+++ backend/src/test/java/kr/momo/controller/schedule/ScheduleControllerTest.java
@@ -1,6 +1,6 @@
 package kr.momo.controller.schedule;
 
-import static kr.momo.service.schedule.ScheduleRecommender.EARLIEST_ORDER;
+import static kr.momo.service.schedule.RecommendedScheduleSortStandard.EARLIEST_ORDER;
 
 import io.restassured.RestAssured;
 import io.restassured.http.ContentType;
